backend/requirements.txt
# =========================
# Runtime target: Python 3.12 (recommended)
# For Python 3.13, some libs (uvloop/httptools/greenlet) are skipped via markers.
# =========================

# FastAPI & Web server
fastapi==0.104.1
uvicorn==0.24.0
uvloop>=0.20.0; python_version < "3.13"      # skip on 3.13
httptools>=0.6.1; python_version < "3.13"    # skip on 3.13
gunicorn==21.2.0
orjson>=3.10.7                               # faster JSON for FastAPI (optional but recommended)

# Database (PostgreSQL-first)
psycopg[binary]==3.2.3
sqlalchemy==2.0.23
alembic==1.12.1
pgvector>=0.2.5                               # SQLAlchemy type for pgvector
# Optional MySQL client (keep only if truly needed)
PyMySQL==1.1.0

# Redis & task queue
redis==5.0.1
celery==5.3.4

# Crawling
playwright>=1.46
beautifulsoup4==4.12.2
requests==2.31.0
lxml>=5.2.1

# Data processing  (Py3.12 recommended for these pins)
pandas==2.2.3
numpy==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0                # .env 파일 로딩 (API 키 관리)
pyyaml==6.0.1
typer>=0.12.0                                 # CLI entrypoints
rich>=13.7.1                                  # nicer CLI logs

# Monitoring & logging
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# Web app / notebooks
streamlit>=1.36.0
jupyter==1.0.0
seaborn==0.13.0

# Dev tools
pytest>=8.2                                     # Py3.13 compatibility
pytest-asyncio==0.21.1
black>=24.8.0
isort==5.12.0
flake8==6.1.0
greenlet>=3.1; python_version < "3.13"          # SQLAlchemy sync-async aid (skip on 3.13)
